requirements.txt
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
pillow==10.2.0
PyJWT==2.6.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
requests
python-dateutil